requirements.txt
aiofiles==0.6.0
amqp==2.6.1
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
boto3==1.16.13
botocore==1.19.63
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colour==0.1.5
constantly==15.1.0
cryptography==3.4.4
cssselect==1.1.0
cycler==0.10.0
DecisionTree==3.4.3
decorator==4.4.2
distlib==0.3.1
dtreeviz==1.2
fastapi==0.64.0
fastapi-jwt-auth==0.3.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
graphviz==0.16
h11==0.12.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
incremental==17.5.0
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jplephem==2.15
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
kombu==4.6.11
lime==0.2.0.1
llvmlite==0.36.0
lunardate==0.2.0
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
minio==4.0.21
mlxtend==0.18.0
mysql-connector==2.2.9
mysql-connector-python==8.0.24
networkx==2.5
numba==0.53.0
numpy==1.20.1
opencv-python-headless==4.5.2.52
packaging==20.9
pandas==1.2.3
parsel==1.6.0
parso==0.8.1
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
Protego==0.1.16
protobuf==3.16.0
py==1.10.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyCalverter==1.6.1
pycparser==2.20
pydantic==1.8.1
PyDispatcher==2.0.5
Pygments==2.8.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyluach==1.2.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-multipart==0.0.5
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
queuelib==1.5.0
requests==2.25.1
s3transfer==0.3.7
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
Scrapy==2.4.1
seaborn==0.11.1
selenium==3.141.0
service-identity==18.1.0
sgp4==2.19
shap @ https://github.com/slundberg/shap/archive/f3369c7.zip
six==1.15.0
skyfield==1.39
skyfield-data==3.0.0
slicer==0.0.7
starlette==0.13.6
threadpoolctl==2.1.0
tifffile==2021.3.31
tensorflow==2.6.0
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
Twisted @ file:///C:/Users/Looi%20Kai%20Wen/Downloads/Twisted-20.3.0-cp39-cp39-win_amd64.whl
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.3
uvicorn==0.13.4
vine==1.3.0
virtualenv==20.4.6
w3lib==1.22.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
workalendar==14.1.0
xgboost==1.3.3
XlsxWriter==1.3.7
yagmail==0.11.224
zope.interface==5.2.0
